fix: rev.1

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.IncorrectParameterException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
@@ -35,21 +36,32 @@ public class FilmService {
     }
 
     public Film likeFilm(Long id, Long userId) {
+        if (id <= 0) {
+            throw new IncorrectParameterException("id");
+        }
+        if (userId <= 0) {
+            throw new IncorrectParameterException("friendId");
+        }
         filmStorage.addLike(id);
         filmStorage.getFilmById(id).getLikes().add(userId);
         return filmStorage.getFilmById(id);
     }
 
     public Film unlikeFilm(Long id, Long userId) {
+        if (id <= 0) {
+            throw new IncorrectParameterException("id");
+        }
+        if (userId <= 0) {
+            throw new IncorrectParameterException("friendId");
+        }
         filmStorage.deleteLike(id);
         filmStorage.getFilmById(id).getLikes().remove(userId);
         return filmStorage.getFilmById(id);
     }
 
     public List<Film> getPopularFilms(Integer count) {
-        return filmStorage.getFilms().stream().sorted((p0, p1) -> {
-            int comp = p0.getCount().compareTo(p1.getCount());
-            return -1 * comp;
-        }).limit(count).collect(Collectors.toList());
+        return filmStorage.getFilms().stream().sorted((p0, p1) -> p1.getCount().compareTo(p0.getCount()))
+                .limit(count)
+                .collect(Collectors.toList());
     }
 }
